requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.0
sqlalchemy>=2.0.36
openpyxl==3.1.2
python-jose[cryptography]==3.3.0
bcrypt==4.0.1
python-dotenv==1.0.1
python-multipart==0.0.9
jinja2==3.1.3
aiofiles==23.2.1
